requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-core
langchain-community
pypdf
requests
pandas
psycopg2-binary
langchain-experimental
tabulate
sqlalchemy